% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_fdr}
\alias{calculate_fdr}
\title{Calculate qvalues of the subset PSMs.}
\usage{
calculate_fdr(df, score_higher = TRUE)
}
\arguments{
\item{df}{A dataframe with at least the columns scores from decoy and subset target PSMs (see details)}

\item{score_higher}{TRUE if higher scores mean a better peptide sequence match.}
}
\value{
A data frame containing the collumns from the original input,
A column with the conservative pi_0 estimation,
the estimated qvalues from the stable FDR estimation strategy using the conservative pi_0 estimator and a large decoy set.
Lastly, a column containing qvalues estimated with the Benjamini Hochbergh procedure.
}
\description{
Calculate qvalues of the subset PSMs.
}

