% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{check_input}
\alias{check_input}
\title{.. content ..}
\usage{
check_input(rawinput, sep = ",")
}
\arguments{
\item{rawinput}{the raw input}

\item{sep}{the separator}
}
\value{

}
\description{
.. content ..
}
\author{

}
\keyword{internal}
