% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{id_is_present}
\alias{id_is_present}
\title{Checks if protein id appears in the headers of a fasta file.}
\usage{
id_is_present(protein_id, fastapath)
}
\arguments{
\item{protein_id}{Vector of protein ids.}

\item{fastapath}{Location of the fasta file.}
}
\value{
Logical vector, TRUE if protein id is present in provided fasta file, FALSE otherwise.
}
\description{
Checks if protein id appears in the headers of a fasta file.
}

