% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pi0plot}
\alias{pi0plot}
\title{Creates density plot of the pi0 distribution}
\usage{
pi0plot(n_targets, n_decoys)
}
\arguments{
\item{n_decoys}{}
}
\description{
Creates density plot of the pi0 distribution
}

