% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_diag}
\alias{plot_diag}
\title{Plot diagnostic plots to evaluate assumptions from the search all, search subset strategy.}
\usage{
plot_diag(df)
}
\arguments{
\item{df}{dataframe with at least 3 columns:
\describe{
\item{score}{score assigned to the peptide to spectrum match (PSM).}
\item{subset}{TRUE if PSM belongs to the subset in interest, FALSE otherwise.}
\item{decoy}{TRUE if decoy PSM, FALSE otherwise.}
}
Additional columns are allowed but ignored.
Target and decoy PSMs are assumbed to be from a competitive target decoy database search.}
}
\description{
Plot diagnostic plots to evaluate assumptions from the search all, search subset strategy.
}

