% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{saas_gui}
\alias{saas_gui}
\title{Launches the GUI version of saas.}
\usage{
saas_gui(port = 3320, host = "0.0.0.0", ...)
}
\arguments{
\item{...}{}
}
\description{
To easily launch the GUI outside an R session (eg. on a server),
you can run R -e "library(saas);saas_gui()" from the terminal (on linux/mac).
}
\details{
see ?shiny::runApp for help all parameters.
}

