% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_fdr}
\alias{calculate_fdr}
\title{Calculates qvalues on the subset PSMs.}
\usage{
calculate_fdr(df, score_higher = TRUE)
}
\arguments{
\item{df}{dataframe with at least 3 columns:
\describe{
\item{score}{score assigned to the peptide to spectrum match (PSM).}
\item{subset}{TRUE if PSM belongs to the subset in interest, FALSE or otherwise.}
\item{decoy}{TRUE if decoy PSM, FALSE otherwise.}
}
Additional columns are allowed but ignored.
Target and decoy PSMs are assumbed to be from a competitive target decoy database search.}

\item{score_higher}{TRUE if a higher score means a better PSM.}
}
\value{
A data frame containing all columns in ``df''.
Following columns are added:
\describe{
\item{pi_0_cons}{conservative estimation of \eqn{\pi_0}.}
\item{FDR}{estimated subset PSM qvalues calculated according the competitive target decoy approach.}
\item{FDR_BH}{estimated subset PSM qvalues calculated according the Benjamini Hochbergh procedure.
When provided, non-subset decoy PSMs are used to stabilize estimates in small subsets}
\item{FDR_stable}{estimated subset PSM qvalues calculated with ``pi_0_cons''.
When provided, non-subset decoy PSMs are used to stabilize estimates in small subsets}
}
}
\description{
Calculates qvalues on the subset PSMs.
}
\examples{

## Simulate a dataset with 140 correct target subset PSMs, 60 incorrect target subset PSMS,
## 60 decoy subset PSMs and 2000 additional decoy PSMs.
set.seed(10)
d = sample_dataset(H1_n = 140,H0_n = 60, decoy_n = 60 ,decoy_large_n = 2000,
                   H0_mean = 2.7, H1_mean = 3, decoy_mean = 2.7, decoy_large_mean = 2.7)

## calculate the qvalues in the subset target PSMS according the classical target-decoy approach
## and our more stable estimation method.
calculate_fdr(d)

}

