% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_msgf_mzid}
\alias{parse_msgf_mzid}
\title{Parses a mzID file generated by MS-GF+.}
\usage{
parse_msgf_mzid(mzid_path)
}
\arguments{
\item{mzid_path}{Location of the mzID file.}
}
\value{
A data frame containing the following 7 columns:
\describe{
\item{spec_id}{Id of the spectrum from the searched dataset file.}
\item{sequence}{Amino acid sequence matching the spectra.}
\item{protein_id}{Id of the sequence from the database file.}
\item{score}{score assigned to the peptide to spectrum match (PSM).}
\item{database}{Name of the database file used to search the spectra.}
\item{decoy}{TRUE if decoy PSM, FALSE otherwise.}
\item{database_size}{Number of sequences in the database file.}
}
}
\description{
See \url{https://omics.pnl.gov/software/ms-gf}
for more info on how to perform a database search on MSMS dataset with MS-GF+ and how to generate a mzID file.
Note that most functions in these package require data from a competitive target decoy search.
}
\details{
We take the MS-GF+ SpecEValue as the PSM score for FDR calculation.
}
\examples{

## Location of the zipped data files
zip_file_path = system.file("extdata", "extdata.zip", package = "saas")

## Unzip and get the (temporary) location of the mzid file with the MS-GF+ search results from a
## competitive target decoy search of the complete pyrococcus proteome against a pyrococcus dataset.
mzid_file_path = unzip(zip_file_path, 'pyrococcus.mzid',exdir = tempdir())
## Parse the mzid file
parse_msgf_mzid(mzid_file_path)

}
