% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_theo_dist}
\alias{plot_theo_dist}
\title{plots the theoretical distribution of all components in the PSM distribution}
\usage{
plot_theo_dist(H0_mean = 2.75, H1_mean = 3.31, H0_sd = 0.13,
  H1_sd = 0.28, decoy_mean = H0_mean, decoy_sd = H0_sd,
  decoy_extra_mean = H0_mean, decoy_extra_sd = H0_sd)
}
\description{
plots the theoretical distribution of all components in the PSM distribution
}

