% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{PPplot}
\alias{PPplot}
\title{Title}
\usage{
PPplot(score, label, pi0 = 0, title = "PP plot of target PSMs",
  xlab = "Decoy percentile", ylab = "Target\\npercentile")
}
\arguments{
\item{ylab}{}
}
\description{
Title
}

