% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess data from a MS-GF mzID file.}
\usage{
preprocess(dat, remove_target_decoy_PSM = TRUE,
  remove_multiple_proteins_PSM = FALSE, is_subset = NULL)
}
\arguments{
\item{dat}{Data frame generated by the saas::parse_msgf_mzid function.}

\item{remove_target_decoy_PSM}{TRUE to remove PSMs that match both a target and decoy sequence.}

\item{remove_multiple_proteins_PSM}{TRUE to remove PSMs that can be assigned to multiple protein ids.}

\item{is_subset}{Location of fasta file with  protein_id of the subset of interest in the fasta headers.}
}
\value{
Data frame with the same columns as ``dat''.
The column protein_id contains all protein_ids that can be assigned to this PSM.
Multiple protein_ids are separated by ``;''.
When ``is_subset'' is specified, two columns are added:
\describe{
\item{subset}{TRUE if sequence can be assigned to a subset protein id}
\item{non_subset}{TRUE if sequence can be assigned to a non subset protein id}
}

Every spectrum haves only 1 row in the data frame.
}
\description{
The parsed data frame from saas::parse_msgf_mzid function contains sometimes multiple entries for a spectrum. (eg. if sequence can be assigned to multiple protein ids). This function takes care of this by default.
}

