% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rpi0}
\alias{rpi0}
\title{Random generation for the \eqn{\pi_0} distribution.}
\usage{
rpi0(n, n_targets, n_decoys)
}
\arguments{
\item{n}{number of observations.}

\item{n_targets}{number of observed target PSMs.}

\item{n_decoys}{number of observed decoy PSMs.}
}
\value{
vector of random deviates.
The length equals `n'.
}
\description{
Random generation for the \eqn{\pi_0} distribution.
}
\examples{

## visualize the pi0 distribution when observing 10 targets and 3 decoys
x = rpi0(100000, 10 ,3 )
hist(x, breaks = 50 , xlab = 'pi0', ylab = 'counts')

}
