% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{saas_gui}
\alias{saas_gui}
\title{Launches the GUI version of saas.}
\usage{
saas_gui(options = list(port = 3320, host = "0.0.0.0"))
}
\arguments{
\item{options}{See help of shiny::shinyApp for more details on available options}
}
\description{
To easily launch the GUI outside an R session (eg. on a server),
you can run R -e "library(saas);saas_gui()" from the terminal (on linux/mac).
}

