% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_dataset}
\alias{sample_dataset}
\title{Sample scores for a random dataset}
\usage{
sample_dataset(H1_n = 160, H0_n = 40, decoy_n = H0_n,
  decoy_large_n = 2000, H0_mean = 2.75, H1_mean = 3.31, H0_sd = 0.13,
  H1_sd = 0.28, decoy_mean = H0_mean, decoy_sd = H0_sd,
  decoy_large_mean = H0_mean, decoy_large_sd = H0_sd)
}
\arguments{
\item{H1_n}{number of correct subset target PSMs.}

\item{H0_n}{number of incorrect subset target PSMs.}

\item{decoy_n}{number of subset decoy PSMs.}

\item{decoy_large_n}{number of non subset decoy PSMs.}

\item{H0_mean}{mean of the incorrect subset target PSM distribution.}

\item{H1_mean}{mean of the correct subset target PSM distribution.}

\item{H0_sd}{sd of the incorrect subset target PSM distribution.}

\item{H1_sd}{sd of the correct subset target PSM distribution.}

\item{decoy_mean}{mean of the subset decoy PSM distribution.}

\item{decoy_sd}{sd of the subset decoy PSM distribution.}

\item{decoy_large_mean}{mean of the non subset decoy PSM distribution.}

\item{decoy_large_sd}{sd of the non subset decoy PSM distribution.}
}
\value{
dataframe with 4 columns:
\describe{
\item{score}{score assigned to the peptide to spectrum match (PSM).}
\item{decoy}{TRUE if decoy PSM, FALSE otherwise.}
\item{H0}{TRUE if incorrect target PSM, FALSE otherwise.}
\item{subset}{TRUE if PSM belongs to the subset in interest, FALSE or otherwise.}
}
}
\description{
The scores are sampled from a two-component mixture distribution of Gaussians.
}
\examples{

## Simulate a dataset with 140 correct target subset PSMs, 60 incorrect target subset PSMS,
##60 decoy subset PSMs and 2000 additional decoy PSMs.
## In this first example, incorrect subset target and decoy PSMs have a similar distribution
set.seed(10)
d = sample_dataset(H1_n = 140,H0_n = 60, decoy_n = 60 ,decoy_large_n = 2000,
                   H0_mean = 2.7, H1_mean = 3, decoy_mean = 2.7, decoy_large_mean = 2.7)
# visualize the decoy distribution and the correct and incorrect target distribution
plot(density(d$score[d$decoy]), xlim = c(2,4), xlab = 'score', ylab = 'density')
lines(density(d$score[!d$decoy & d$H0]), col = 'red')
lines(density(d$score[!d$decoy & !d$H0]), col = 'green')
legend('topright', c('decoys', 'incorrect targets', 'correct targets'),
       text.col = c('black', 'red', 'green'))
## In this first example, incorrect subset target and decoy PSMs have not a similar distribution
## because the additional set of decoy PSMs are not representative for the incorrect subset PSMs.
set.seed(10)
d = sample_dataset(H1_n = 140,H0_n = 60, decoy_n = 60 ,decoy_large_n = 2000,
                   H0_mean = 2.7, H1_mean = 3, decoy_mean = 2.7, decoy_large_mean = 2.6)
# visualize the decoy distribution and the correct and incorrect target distribution
plot(density(d$score[d$decoy]), xlim = c(2,4), xlab = 'score', ylab = 'density')
lines(density(d$score[!d$decoy & d$H0]), col = 'red')
lines(density(d$score[!d$decoy & !d$H0]), col = 'green')
legend('topright', c('decoys', 'incorrect targets', 'correct targets'),
       text.col = c('black', 'red', 'green'))

}
\keyword{internal}
